backend/requirements.txt
adium-theme-ubuntu==0.3.4
appdirs==1.4.0
beautifulsoup4==4.5.3
biopython==1.68
bioservices==1.4.14
certifi==2018.10.15
chardet==3.0.4
click==6.7
click-conf==0.0.1
click-stream==0.0.6
colorama==0.3.7
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
demjson==2.2.4
diff-match-patch==20121119
dj-static==0.0.6
Django==1.11.16
django-crispy-forms==1.6.1
django-filter==1.0.1
django-import-export==0.5.1
django-ipware==1.1.6
django-rest-swagger==2.2.0
djangorestframework==3.9.1
docutils==0.14
easydev==0.9.30
elasticsearch==6.1.1
elasticsearch-loader==0.2.24
future==0.18.2
gevent==1.2.0
goatools==0.7.9
greenlet==0.4.11
grequests==0.3.0
gunicorn==19.10.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.6
MarkupSafe==1.1.0
more-itertools==4.3.0
names==0.3.0
nose==1.3.7
numpy==1.11.3
openapi-codec==1.3.2
pandas==0.20.3
patsy==0.4.1
pexpect==4.2.1
psutil==5.4.5
ptyprocess==0.5.1
pyaml==18.11.0
pycrypto==2.6.1
pygobject==3.20.0
pymssql==2.1.4
python-dateutil==2.6.1
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-cache==0.4.13
rpy2==2.7.8
schedule==0.5.0
scipy==0.19.1
scour==0.32
selenium==3.0.1
sh==1.13.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
static3==0.7.0
statistics==1.0.3.5
statsmodels==0.8.0
suds-jurko==0.6
tablib==0.11.2
unicodecsv==0.14.1
unity-lens-photos==1.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.0.0
wget==3.2
wrapt==1.10.8
xlrd==1.1.0
XlsxWriter==1.0.0
xmltodict==0.10.2
